+++ package.json
@@ -19,7 +19,8 @@
     "jwt-decode": "^4.0.0",
     "notistack": "^3.0.2",
     "react": "^19.1.1",
-    "react-dom": "^19.1.1"
+    "react-dom": "^19.1.1",
+    "zustand": "^5.0.8"
   },
   "devDependencies": {
     "@eslint/js": "^9.33.0",
