+++ src/entities/Todo/ui/Todos.tsx
@@ -1,108 +1,21 @@
-import { Card, CardActions, CardContent, Checkbox, Stack, TextField } from '@mui/material'
-import Typography from '@mui/material/Typography'
+import { Container, Input, Stack } from '@mui/material'
+import { useTodosStore } from '../model/store/use TodosStore.ts'
+import { Todo } from './Todo.tsx'
+import { type ChangeEvent, useState } from 'react'
+import Button from '@mui/material/Button'
 import type { TodoType } from '../model/todoType.ts'
-import { mockTodos } from '../model/mockTodos.ts'
-import { type SyntheticEvent, useState } from 'react'
-import IconButton from '@mui/material/IconButton'
-import ModeEditIcon from '@mui/icons-material/ModeEdit'
-import DoneIcon from '@mui/icons-material/Done'
-import { useSnackbar } from 'notistack'
-
-type TodoProps = {
-	todo: TodoType
-	setTodo: (todo: TodoType) => void
-}
-
-const Todo = ({ todo, setTodo }: TodoProps) => {
-	const [isTitle, setIsTitle] = useState(true)
-	const [isDescription, setIsDescription] = useState(true)
-
-	const [editedTitle, setEditedTitle] = useState(todo.title)
-	const [editedDescription, setEditedDescription] = useState(todo.description)
-
-	const handleCheckClick = () => {
-		setTodo({ ...todo, completed: !todo.completed })
-	}
-
-	const handleTitleChanger = () => {
-		setIsTitle(!isTitle)
-	}
-
-	const handleTitleChangerTextField = (e: SyntheticEvent<HTMLTextAreaElement | HTMLInputElement>) => {
-		const newValue = e.currentTarget.value
-		setEditedTitle(newValue)
-		setTodo({ ...todo, title: newValue, updatedAt: Date() })
-	}
-
-	const handleDescriptionChanger = () => {
-		setIsDescription(!isDescription)
-	}
-
-	const handleTitleChangerDescriptionTextField = (e: SyntheticEvent<HTMLTextAreaElement | HTMLInputElement>) => {
-		const newValue = e.currentTarget.value
-		setEditedDescription(newValue)
-		setTodo({ ...todo, description: newValue, updatedAt: Date() })
-	}
-
-	const DoneEditTitle = () => {
-		setIsTitle(!isTitle)
-		enqueueSnackbar('Заголовок успешно обновлен!', { variant: 'success' })
-	}
-
-	const DoneEditDescription = () => {
-		setIsDescription(!isDescription)
-		enqueueSnackbar('Описание успешно обновлено!', { variant: 'success' })
-	}
-
-	const { enqueueSnackbar } = useSnackbar()
-
-	return (
-		<Card variant="outlined" sx={{ maxWidth: 200 }}>
-			<CardContent>
-				{isTitle ? (
-					<>
-						<Typography gutterBottom sx={{ color: 'text.secondary', fontSize: 14 }}>
-							{todo.title}
-						</Typography>
-						<IconButton size="small" onClick={handleTitleChanger}>
-							<ModeEditIcon fontSize="small" />
-						</IconButton>
-					</>
-				) : (
-					<>
-						<TextField label={editedTitle} onChange={handleTitleChangerTextField}></TextField>
-						<IconButton size="small" onClick={DoneEditTitle}>
-							<DoneIcon fontSize="small" />
-						</IconButton>
-					</>
-				)}
-				{isDescription ? (
-					<>
-						<Typography variant="body2">{todo.description}</Typography>
-						<IconButton size="small" onClick={handleDescriptionChanger}>
-							<ModeEditIcon fontSize="small" />
-						</IconButton>
-					</>
-				) : (
-					<>
-						<TextField label={editedDescription} onChange={handleTitleChangerDescriptionTextField}></TextField>
-						<IconButton size="small" onClick={DoneEditDescription}>
-							<DoneIcon fontSize="small" />
-						</IconButton>
-					</>
-				)}
-				<Typography>{`Дата создания: ${new Date(todo.createdAt).toLocaleDateString()}`}</Typography>
-				<Typography>{`Дата обновления: ${new Date(todo.updatedAt).toLocaleDateString()}`}</Typography>
-			</CardContent>
-			<CardActions>
-				<Checkbox checked={todo.completed} onClick={handleCheckClick}></Checkbox>
-			</CardActions>
-		</Card>
-	)
-}
 
 const Todos = () => {
-	const [todos, setTodos] = useState<TodoType[]>(mockTodos)
+	//	const [todos, setTodos] = useState<TodoType[]>(mockTodos)
+
+	const [newTodoTitle, setNewTodoTitle] = useState('')
+	const [newTodoDescription, setNewTodoDescription] = useState('')
+	const todos = useTodosStore((state) => state.todos)
+	const addTodos = useTodosStore((state) => state.addTodo)
+	const setTodos = useTodosStore((state) => state.setTodos)
+
+	console.log(todos)
+
 	const setTodo = (todo: TodoType) => {
 		const updatedTodos = todos.map((t) => {
 			if (t._id === todo._id) {
@@ -113,12 +26,41 @@ const Todos = () => {
 		setTodos(updatedTodos)
 	}
 
+	const handleTitleChange = (e: ChangeEvent<HTMLInputElement>) => {
+		setNewTodoTitle(e.target.value)
+	}
+
+	const handleDescriptionChange = (e: ChangeEvent<HTMLInputElement>) => {
+		setNewTodoDescription(e.target.value)
+	}
+
+	const handleAddTodo = () => {
+		const newTodo: TodoType = {
+			_id: Date.now().toString(),
+			title: newTodoTitle,
+			description: newTodoDescription,
+			completed: false,
+			createdAt: new Date().toISOString(),
+			updatedAt: new Date().toISOString(),
+			order: todos.length + 1,
+		}
+		addTodos(newTodo)
+	}
+
 	return (
-		<Stack flexWrap={'wrap'} spacing={2} direction={'row'}>
-			{todos.map((todo) => {
-				return <Todo todo={todo} key={todo._id} setTodo={setTodo} />
-			})}
-		</Stack>
+		<Container>
+			<Input placeholder={'title'} value={newTodoTitle} onChange={handleTitleChange} />
+			<Input placeholder={'description'} value={newTodoDescription} onChange={handleDescriptionChange} />
+			<Button disabled={!newTodoTitle} onClick={handleAddTodo}>
+				Add
+			</Button>
+
+			<Stack flexWrap={'wrap'} spacing={2} direction={'row'}>
+				{todos.map((todo) => {
+					return <Todo todo={todo} key={todo._id} setTodo={setTodo} />
+				})}
+			</Stack>
+		</Container>
 	)
 }
